+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.11.11 (07-01-2021)
 * Bugfixes
+* Wrong method parameter JS Linter added
 
 ## 0.11.10 (07-01-2021)
 * Wrong SAP UI Define import linting added